Refactor into separate checking script.
